requirements.txt
gunicorn==20.0.4
Brotli==1.0.9
click==7.1.2
cssmin==0.2.0
dash==1.15.0
dash-core-components==1.11.0
dash-html-components==1.1.0
dash-bootstrap-components
dash-renderer==1.7.0
dash-table==4.10.0
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
lesscpy==0.14.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
pathlib==1.0.1
plotly==4.9.0
ply==3.11
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
retrying==1.3.3
requests_toolbelt==0.9.1
six==1.15.0
webassets==2.0
Werkzeug==1.0.1
xlrd==1.1.0
